Discard interface idea since we’re only going to have one implementation (now and in the foreseeable future)
